+++ SFA.DFC.TestAutomation/SFA.DFC.FindACourse.UITests/Project/Tests/Features/ApplyFiltersForCourseSearchResults.feature
@@ -6,8 +6,8 @@
 @Findacourse
 Scenario: Search for course and Apply filters
 	Given I have navigated to Find a course page
-	And I have searched for a  course 'Maths'
-	When I click the Find a course button
+	When I have searched for a  course 'Maths'
+	And I click the Find a course button
 	Then the results for the course should be listed	
 	When I apply the filter hours 'Full time', type 'Online' and start date 'From Today'
 	And I have clicked the Apply Filter button
