+++ SFA.DFC.TestAutomation/SFA.DFC.FindACourse.UITests/Project/Tests/Features/SearchForACourseWithOtherDetails.feature
@@ -4,12 +4,12 @@
 @Findacourse
 Scenario: Search for course and other information
 	Given I have navigated to Find a course page
-	And I have searched for a  course 'Maths'
+	When I have searched for a  course 'Maths'
 	And I have entered a provider 'Leicester College' and location 'Leicester' 
 	When I click the Find a course button
 	Then the results for the course should be listed	
 	When I change the provider 'Skills' and location 'Birmingham'
 	And I have clicked the Apply Filter button
-	Then the results for the new provider 'Skills' and location 'Birmingham' should be displayed
+	Then the results for the new provider and location should be displayed
